Split `FilterPublicMembershipChangesPostProcessor` from `RoomBeginningPostProcessor`
