+++ libraries/matrix/impl/src/main/kotlin/io/element/android/libraries/matrix/impl/timeline/postprocessor/RoomBeginningPostProcessor.kt
@@ -9,12 +9,10 @@
 package io.element.android.libraries.matrix.impl.timeline.postprocessor
 
 import io.element.android.libraries.matrix.api.core.UserId
-import io.element.android.libraries.matrix.api.room.join.JoinRule
 import io.element.android.libraries.matrix.api.timeline.MatrixTimelineItem
 import io.element.android.libraries.matrix.api.timeline.Timeline
 import io.element.android.libraries.matrix.api.timeline.item.event.MembershipChange
 import io.element.android.libraries.matrix.api.timeline.item.event.OtherState
-import io.element.android.libraries.matrix.api.timeline.item.event.ProfileChangeContent
 import io.element.android.libraries.matrix.api.timeline.item.event.RoomMembershipContent
 import io.element.android.libraries.matrix.api.timeline.item.event.StateContent
 
@@ -27,28 +25,14 @@ class RoomBeginningPostProcessor(private val mode: Timeline.Mode) {
         items: List<MatrixTimelineItem>,
         isDm: Boolean,
         roomCreator: UserId?,
-        joinRule: JoinRule?,
-        isEncrypted: Boolean?,
         hasMoreToLoadBackwards: Boolean,
     ): List<MatrixTimelineItem> {
         return when {
             items.isEmpty() -> items
             mode == Timeline.Mode.PinnedEvents -> items
-            joinRule !is JoinRule.Invite && isEncrypted == false -> filterRoomMemberEvents(items)
             isDm -> processForDM(items, roomCreator)
             hasMoreToLoadBackwards -> items
             else -> processForRoom(items)
-        }
-    }
-
-    private fun filterRoomMemberEvents(items: List<MatrixTimelineItem>): List<MatrixTimelineItem> {
-        return items.filter { item ->
-            val eventContent = (item as? MatrixTimelineItem.Event)?.event?.content
-            when (eventContent) {
-                is RoomMembershipContent -> eventContent.change !in listOf(MembershipChange.JOINED, MembershipChange.LEFT)
-                is ProfileChangeContent -> false
-                else -> true
-            }
         }
     }
